in which julian adds strategies, and realizes that they aren't helping that much

+++ src/main/scala/com/diatom/agent/TDeletorAgent.scala
@@ -3,24 +3,44 @@ package com.diatom.agent
 import akka.actor.{ActorRef, ActorSystem, Props}
 import com.diatom.agent.func.{TDeletorFunc, TMutatorFunc}
 import com.diatom.population.TPopulation
+import scala.concurrent.duration._
 
 trait TDeletorAgent[Sol] extends TAgent[Sol]
 
-case class DeletorAgent[Sol](delete: TDeletorFunc[Sol], pop: TPopulation[Sol])
-  extends AAgent[Sol] with TDeletorAgent[Sol] {
+case class DeletorAgent[Sol](delete: TDeletorFunc[Sol], pop: TPopulation[Sol], strat: TAgentStrategy)
+  extends AAgent[Sol](strat, pop) with TDeletorAgent[Sol] {
 
   override protected def step(): Unit = {
     val in = pop.getSolutions(delete.numInputs)
-    val toDelete = delete.delete(in)
-    pop.deleteSolutions(toDelete)
+    if (in.size == delete.numInputs) {
+      val toDelete = delete.delete(in)
+      pop.deleteSolutions(toDelete)
+    }
   }
 }
 
 object DeletorAgent {
+
   def from[Sol](deletorFunc: TDeletorFunc[Sol], pop: TPopulation[Sol])
                (implicit system: ActorSystem): ActorRef = {
-    system.actorOf(Props(DeletorAgent(deletorFunc, pop)))
+    val strat: TAgentStrategy = _ => 50.millis // TODO replace with sane default
+    DeletorAgent.from(deletorFunc, pop, strat)
+  }
+
+  def from[Sol](deletorFunc: TDeletorFunc[Sol], pop: TPopulation[Sol], strat: TAgentStrategy)
+               (implicit system: ActorSystem): ActorRef = {
+    system.actorOf(Props(DeletorAgent(deletorFunc, pop, strat)))
   }
 }
 
 
+case class DeletorAgentDefaultStrategy() extends TAgentStrategy {
+  override def waitTime(populationInformation: TPopulationInformation): Duration = {
+    if (populationInformation.numSolutions < 20) {
+      // min of 1 and fourth root of num solutions. No particular reason why.
+      200.millis // give creators a chance!
+    } else {
+      math.max(1, math.sqrt(math.sqrt(populationInformation.numSolutions))).millis
+    }
+  }
+}
